[Update] reset dialogue 및 chat 디자인 수정
변경된 Figma 디자인에 맞춰 chat 디자인 수정
